@@ -25,7 +25,7 @@ import vmnet
 
 #if !CURRENT_SDK
 /// Interface strategy for containers that use macOS's custom network feature.
-@available(macOS 16, *)
+@available(macOS 26, *)
 struct NonisolatedInterfaceStrategy: InterfaceStrategy {
     private let log: Logger
 
